Add methods
I added two methods, one to make a pile of rocks and the other to make a well. I also made a logic statement that chooses whether to display the well or the pile of rocks. Finally, I edited the constructor method to have a "this.smlPxl" and a "this.well" as well.