Change certain "TextBox"es to "CollectionBox"es
I changed certain "TextBox"es to "CollectionBox"es. I also made it so that the map worked with key presses and not button presses. I added in an explanation as to how the map worked. I made a "CollectionBox" for when you collect the ruby. I also fixed Janine's ring quest, because you didn't need to collect the ring to complete the quest before. Finally, I added something to the large comment at the top.